Move the helpers to their conventional location
